Refactored IRag methods that should have been in IGenerellemIngestion.
